feat: @silent support
Discord [relatively] recently added support for supressing notifications by typing @silent before the message content. This applies 1 << 12 to the message flags, and does not emit a notification, but still shows the highlight/mention in the channel you've been mentioned in.

Up until now, this bot did not support this functionality, neither in the sense of utilizing the feature or respecting messages attributed with 1 << 12.

This commit is the main purpose of the subsequent PR this will be in, allowing for users to quietly mention users in a channel, effectively bringing the bot in parity with the original message in that sense as well.